third_party/WebKit/LayoutTests/fast/multicol/explicit-columns-auto.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
    .explicitlyNoColumns { -webkit-columns:auto; }
</style>

<div id="elm"></div>

<script>
    description("Test that setting columns to auto is the same as not setting it at all.");
    document.body.offsetTop;
    document.getElementById('elm').className = "explicitlyNoColumns";
    if (window.internals) {
        // Need to call updateStyleAndReturnAffectedElementCount() first, to get style recalculated,
        // so that the necessary objects get marked for layout (which, if the test passes, should be 0).
        shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");
        shouldBe("internals.needsLayoutCount()", "0");
    } else {
        testFailed("There's no web exposed API usable for this test. The bug doesn't affect layout in a reliable way. Please rerun inside the layout test framework.");
    }
</script>
