third_party/blink/web_tests/platform/generic/fast/multicol/explicit-columns-auto-expected.txt

Test that setting columns to auto is the same as not setting it at all.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.needsLayoutCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
